requirements.txt
opencv-python~=4.13.0.92
numpy~=2.4.4
vidgear~=0.3.4
PyYAML~=6.0.3
decord~=0.6.0
scikit-image~=0.26.0